extract password and create PasswordField component
